# docs/02-api/openapi.yaml
openapi: 3.0.3
info:
  title: HK Small Store API Contract
  version: 1.0.0
  license:
    name: MIT
  description: |
    API Contract chuẩn hóa cho hệ thống cửa hàng đồ si tuyển chọn độc bản (1-of-1) HK Small Store.
    Tài liệu này đóng vai trò Single Source of Truth (SSOT) cho việc phát triển Frontend độc lập với Mock Service (MSW) và Backend sau này.

    Quy chuẩn nghiệp vụ cốt lõi:
    - Mỗi món hàng là độc bản duy nhất (1-of-1), số lượng tồn kho = 1.
    - Phân loại chi tiết sản phẩm: `t_shirts`, `shirts`, `sweaters`, `jackets`, `blazers`, `pants`, `shorts`, `skirts`, `dresses`, `bags`, `scarves`, `hats`, `accessories`.
    - 5 Cấp độ tình trạng chuẩn hóa: `new`, `like_new`, `good`, `fair`, `attention_required`.
    - Tuyệt đối không dùng điểm lẻ (9.5/10), thang phần trăm hoặc mã SKU (dùng Mã món - `item_id`).
    - 6 Trạng thái đơn hàng: `new`, `confirmed`, `shipping`, `completed`, `returned`, `cancelled`.
    - 3 Trạng thái thu tiền: `unpaid`, `pending_cod`, `paid`.
    - Tiền tệ: Đồng Việt Nam (VND, số nguyên integer).
servers:
  - url: http://localhost:3000/api
    description: Mock API Server / Local Development
paths:
...
  /admin/admins:
    get:
      operationId: listAdmins
      summary: Danh sách tài khoản admin
      description: Cần quyền `manage_admins` (owner luôn có).
      tags:
        - Admin Accounts
      security:
        - AdminAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [admins]
                properties:
                  admins:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createAdmin
      summary: Tạo tài khoản admin
      description: |
        Cần `manage_admins`. Chỉ **owner** mới được tạo role `owner`.
        Staff nhận `permissions` checkbox; owner bỏ qua permissions (đủ quyền).
      tags:
        - Admin Accounts
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username trùng
  /admin/admins/{id}:
    patch:
      operationId: updateAdmin
      summary: Cập nhật tài khoản admin
      description: Cần `manage_admins`. Không thể hạ/vô hiệu hóa owner active cuối cùng. Chỉ owner gán role owner.
      tags:
        - Admin Accounts
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # ADMIN ITEMS MANAGEMENT
  # ==========================================
  /admin/items:
    get:
      operationId: getAdminItems
      summary: Danh sách quản lý tất cả món hàng trong kho
      description: Dành cho admin, xem được cả đồ nháp, đang giữ chỗ, đã bán và lưu kho.
      tags:
        - Admin Items
      security:
        - AdminAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemStatus'
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemCategory'
        - name: condition
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemCondition'
        - name: batch_id
          in: query
          required: false
          schema:
            type: string
          description: Lọc theo mã kiện hàng nhập về
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest, price_asc, price_desc, name_asc, name_desc]
            default: newest
          description: Sắp xếp theo ngày tạo, giá hoặc tên (trước khi phân trang)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Danh sách món đồ quản trị kèm thống kê nhanh số lượng
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - pagination
                  - stats
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemAdminSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  stats:
                    type: object
                    required:
                      - total_items
                      - on_shelf
                      - reserved
                      - sold
                      - draft
                    properties:
                      total_items:
                        type: integer
                        example: 248
                      on_shelf:
                        type: integer
                        example: 180
                      reserved:
                        type: integer
                        example: 12
                      sold:
                        type: integer
                        example: 44
                      draft:
                        type: integer
                        example: 12
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createAdminItem
      summary: Tạo mới một món đồ lưu trữ
      description: |
        Nhập thông tin món hàng. Nếu đầy đủ ảnh, kích thước và phân loại thì có thể đưa thẳng lên kệ (`shelf`), ngược lại lưu dạng nháp (`draft`).
        Nếu tình trạng là `attention_required`, bắt buộc phải có mô tả vết lỗi và ít nhất 1 ảnh chụp cận cảnh khuyết điểm.
      tags:
        - Admin Items
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpsertRequest'
      responses:
        '201':
          description: Tạo món đồ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemAdminDetail'
        '400':
          description: Dữ liệu không hợp lệ. Lưu thẳng lên kệ mà thiếu điều kiện trả `PUBLISH_REQUIREMENTS_NOT_MET` kèm `fields` (ví dụ thiếu ảnh lỗi khi chọn 'Cần lưu ý').
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/orders:
    get:
      operationId: getAdminOrders
      summary: Danh sách đơn hàng quản trị
      description: Hiển thị đơn hàng với đầy đủ bộ lọc 6 trạng thái xử lý đơn và 3 trạng thái thu tiền.
      tags:
        - Admin Orders
      security:
        - AdminAuth: []
      parameters:
        - name: order_status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: payment_status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: carrier_name
          in: query
          required: false
          schema:
            type: string
            description: Đơn vị vận chuyển (GHN, GHTK, ViettelPost)
        - name: search
          in: query
          required: false
          schema:
            type: string
            description: Tìm kiếm theo mã đơn, SĐT hoặc tên khách
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest, total_desc, total_asc]
            default: newest
          description: Sắp xếp theo ngày đặt hoặc tổng tiền (trước khi phân trang)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Danh sách đơn hàng và số lượng theo từng trạng thái
          content:
            application/json:
              schema:
                type: object
                required:
                  - orders
                  - pagination
                  - counts
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderAdminSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  counts:
                    type: object
                    required:
                      - all
                      - new
                      - confirmed
                      - shipping
                      - completed
                      - returned
                      - cancelled
                    properties:
                      all:
                        type: integer
                        example: 142
                      new:
                        type: integer
                        example: 3
                      confirmed:
                        type: integer
                        example: 5
                      shipping:
                        type: integer
                        example: 8
                      completed:
                        type: integer
                        example: 120
                      returned:
                        type: integer
                        example: 4
                      cancelled:
                        type: integer
                        example: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CustomerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token phiên khách hàng (trường `token` khi đăng nhập), cho client API và kiểm thử. Web dùng cookie `CustomerSessionCookie`.
    CustomerSessionCookie:
      type: apiKey
      in: cookie
      name: hk_small_store_customer_session
      description: |
        Phiên khách hàng của web (ADR 006): cookie httpOnly, `SameSite=Lax`, `Secure` ở production, `Path=/`, `Max-Age=34560000` (400 ngày, mức tối đa trình duyệt cho phép).
        Gia hạn trượt: mỗi lần khách dùng web, cookie được cấp lại thêm 400 ngày. Không có hết hạn theo phiên trình duyệt hay theo thời gian không hoạt động; chỉ hết khi khách bấm Đăng xuất (thiết bị đó), Đăng xuất khỏi mọi thiết bị, đổi mật khẩu (các thiết bị khác) hoặc 400 ngày không vào web.
  responses:
    Unauthorized:
      description: Không có quyền truy cập hoặc phiên làm việc hết hạn
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Không tìm thấy tài nguyên yêu cầu
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Yêu cầu không hợp lệ (`fields` chỉ ra ô cần sửa nếu có)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Đã đăng nhập nhưng thiếu quyền (RBAC)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ItemCategory:
      type: string
      description: |
        Phân loại chi tiết sản phẩm:
        - `t_shirts`: Áo thun / Polo
        - `shirts`: Áo sơ mi
        - `sweaters`: Áo len
        - `jackets`: Áo khoác
        - `blazers`: Blazer / Măng tô
        - `pants`: Quần dài (Jeans, Kaki, Âu)
        - `shorts`: Quần short
        - `skirts`: Chân váy
        - `dresses`: Đầm
        - `bags`: Túi xách / Balo
        - `scarves`: Khăn
        - `hats`: Mũ / Nón
        - `accessories`: Phụ kiện khác
      enum:
        - t_shirts
        - shirts
        - sweaters
        - jackets
        - blazers
        - pants
        - shorts
        - skirts
        - dresses
        - bags
        - scarves
        - hats
        - accessories
    ItemCondition:
      type: string
      description: |
        5 Cấp độ tình trạng chuẩn hóa:
        - `new`: Mới (Chưa qua sử dụng, còn tem mác nguyên bản)
        - `like_new`: Như mới (Gần như mới, phom dáng và chất vải nguyên vẹn)
        - `good`: Tốt (Đã qua sử dụng nhưng giữ gìn rất tốt)
        - `fair`: Đã qua sử dụng (Có độ mòn thời gian tự nhiên, không rách hỏng lớn)
        - `attention_required`: Cần lưu ý (Có khuyết điểm, vết ố, sờn chỉ - có ảnh chụp cận cảnh và mô tả)
      enum: [new, like_new, good, fair, attention_required]
    ItemStatus:
      type: string
      description: |
        Vòng đời trạng thái món đồ:
        - `draft`: Bản nháp (Đang xử lý đo đạc/chụp ảnh, chưa lên kệ)
        - `shelf`: Đang trên kệ (Khách có thể xem và đặt mua)
        - `reserved`: Đang giữ chỗ (Đã nằm trong đơn hàng chờ xử lý)
        - `sold`: Đã bán (Đơn hàng đã hoàn tất)
        (`archived` đã bỏ theo ADR 003 D4.)
      enum: [draft, shelf, reserved, sold]
    OrderStatus:
      type: string
      description: |
        6 Trạng thái quy trình đơn hàng:
        - `new`: Đơn mới tạo từ website
        - `confirmed`: Đã xác nhận & chốt phí ship với khách
        - `shipping`: Đã bàn giao bưu cục kèm mã vận đơn
        - `completed`: Giao thành công cho khách
        - `returned`: Khách trả hàng trong 2 ngày kể từ hoàn tất (chỉ trả, không đổi)
        - `cancelled`: Đơn bị hủy
      enum: [new, confirmed, shipping, completed, returned, cancelled]
    PaymentMethod:
      type: string
      enum: [cod, bank_transfer]
    PaymentStatus:
      type: string
      description: |
        3 Trạng thái thanh toán:
        - `unpaid`: Chưa thanh toán
        - `pending_cod`: Chờ đối soát tiền thu hộ COD từ bưu cục
        - `paid`: Đã nhận tiền vào tài khoản
      enum: [unpaid, pending_cod, paid]
...
    ItemImage:
      type: object
      description: Một ảnh của món kèm chữ thay thế (alt) bắt buộc khi lên kệ (ADR 003 D11).
      required:
        - url
        - alt
      properties:
        url:
          type: string
          example: "https://images.unsplash.com/photo-1598033129183-c4f50c736f10"
        alt:
          type: string
          example: "Áo sơ mi lụa, mặt trước"
    ItemPublicSummary:
      type: object
      required:
        - id
        - name
        - category
        - condition
        - price
        - size
        - main_image
        - status
      properties:
        id:
          type: string
          example: "KN2410-001"
        name:
          type: string
          example: "Áo sơ mi lụa tơ tằm cổ điển 1990s"
        category:
          $ref: '#/components/schemas/ItemCategory'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        price:
          type: integer
          example: 320000
        size:
          type: string
          example: "M (Form suông)"
        main_image:
          type: string
          description: URL ảnh chính (bằng `images[0].url`)
          example: "https://images.unsplash.com/photo-1598033129183-c4f50c736f10"
        status:
          $ref: '#/components/schemas/ItemStatus'
        batch_id:
          type: string
          example: "KN2410"
...
    ItemAdminSummary:
      allOf:
        - $ref: '#/components/schemas/ItemPublicSummary'
        - type: object
          required:
            - batch_id
            - created_at
          properties:
            batch_id:
              type: string
              example: "KN2410"
            created_at:
              type: string
              format: date-time
    ItemAdminDetail:
      allOf:
        - $ref: '#/components/schemas/ItemPublicDetail'
        - type: object
          required:
            - batch_id
            - status
            - created_at
            - updated_at
            - cost_price
          properties:
            batch_id:
              type: string
              example: "KN2410"
            cost_price:
              type: integer
              nullable: true
              description: Giá vốn của món (VND). `null` khi tính theo kiện (ADR 002).
              example: 120000
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    ItemUpsertRequest:
      type: object
      required:
        - name
        - batch_id
        - category
        - condition
        - price
        - size
        - material
        - measurements
        - images
        - status
      properties:
        name:
          type: string
          example: "Áo sơ mi lụa tơ tằm cổ điển"
        batch_id:
          type: string
          example: "KN2410"
        category:
          $ref: '#/components/schemas/ItemCategory'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        price:
          type: integer
          minimum: 0
          example: 320000
        cost_price:
          type: integer
          nullable: true
          minimum: 0
          description: Giá vốn (VND). Bỏ trống nếu tính theo kiện.
          example: 120000
        size:
          type: string
          example: "M"
        material:
          type: string
          example: "100% Lụa Mulberry"
        origin:
          type: string
          example: "Nhật Bản"
        measurements:
          $ref: '#/components/schemas/ItemMeasurements'
        images:
          type: array
          description: Ảnh theo thứ tự hiển thị, ảnh đầu là ảnh chính. Lên kệ cần ≥1 ảnh có `alt`. Nháp có thể rỗng.
          items:
            $ref: '#/components/schemas/ItemImage'
        defect_description:
          type: string
          nullable: true
          description: Bắt buộc nếu condition là attention_required
        defect_images:
          type: array
          items:
            $ref: '#/components/schemas/ItemImage'
          description: Bắt buộc (≥1) nếu condition là attention_required
        status:
          type: string
          enum: [draft, shelf]
          description: Lưu nháp (`draft`) hoặc lưu và lên kệ (`shelf`)
...
    OrderAdminSummary:
      type: object
      required:
        - default_shipping_fee
        - freeship_applied
        - deposit_status
        - deposit_amount
        - hold_expires_at
        - amount_due
        - id
        - order_code
        - customer_name
        - customer_phone
        - total_items
        - subtotal
        - payment_method
        - order_status
        - payment_status
        - created_at
      properties:
        default_shipping_fee:
          type: integer
          nullable: true
          description: Phí ship mặc định (1 chiều) lấy từ cài đặt `default_shipping_fee` lúc đặt đơn, trước khi tính freeship. null với đơn tạo trước khi có phí mặc định.
          example: 30000
        freeship_applied:
          type: boolean
          description: Đơn đủ số món để freeship (cài đặt `freeship_min_items`, chốt lúc đặt) nên `shipping_fee` lúc đặt = 0. Shop vẫn đổi được phí khi xác nhận (khu vực đặc biệt); UI admin cảnh báo khi đổi.
          example: false
        id:
          type: string
          example: "order_123"
        order_code:
          type: string
          example: "DH-260901"
        customer_name:
          type: string
          example: "Nguyễn Văn An"
        customer_phone:
          type: string
          example: "0912345678"
        total_items:
          type: integer
          example: 2
        subtotal:
          type: integer
          example: 770000
        shipping_fee:
          type: integer
          nullable: true
          example: 30000
        total:
          type: integer
          example: 800000
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        carrier_name:
          type: string
          nullable: true
          example: "GHN"
        deposit_status:
          $ref: '#/components/schemas/DepositStatus'
        deposit_amount:
          type: integer
          description: Tiền cọc của đơn, chốt theo cài đặt lúc tạo đơn (mặc định 50.000₫ với số điện thoại chưa có đơn hoàn tất, 0 nếu không cần cọc)
          example: 50000
        amount_due:
          type: integer
          description: Số tiền khách còn phải trả cho đơn = `total` − tiền cọc đã nhận (cọc đã chuyển được trừ vào đơn)
          example: 750000
        hold_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Hết giờ giữ đơn (đơn `new`). Hết giờ mà shop chưa xử lý → đơn cần cọc chưa nhận cọc thì tự hủy (món về kệ); đơn không cần cọc hoặc đã nhận cọc thì tự động xác nhận. Chỉ có ý nghĩa khi đơn còn `new`.
        tracking_code:
          type: string
          nullable: true
          example: "GHN88291039VN"
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required:
        - total
        - page
        - limit
        - total_pages
      properties:
        total:
          type: integer
          example: 120
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 12
        total_pages:
          type: integer
          example: 10
    ErrorResponse:
      type: object
      description: |
        Mã lỗi thường gặp: `INVALID_TRANSITION`, `RETURN_WINDOW_EXPIRED`, `DEPOSIT_REQUIRED`, `HOLD_ALREADY_EXTENDED`, `CANCEL_NOT_ALLOWED`, `ORDER_STATUS_CHANGED`, `POLICY_NOT_ACCEPTED`, `POLICY_VERSION_OUTDATED`, `SHIPPING_FEE_REQUIRED`, `PUBLISH_REQUIREMENTS_NOT_MET`,
        `ITEM_IN_ACTIVE_ORDER`, `EMAIL_NOT_VERIFIED`, `INVALID_CREDENTIALS`, `UNAUTHORIZED`, `FORBIDDEN`, `VALIDATION_FAILED`, `NOT_FOUND`.
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "ITEM_NOT_FOUND"
        message:
          type: string
          example: "Món hàng không tồn tại hoặc đã rời khỏi kệ."
        fields:
          type: array
          description: Lỗi theo từng ô để giao diện hiện ngay dưới ô
          items:
            $ref: '#/components/schemas/FieldErrorDetail'
        email:
          type: string
          format: email
          description: Có khi `EMAIL_NOT_VERIFIED`, để chuyển sang bước nhập OTP
    AdminPermission:
      type: string
      enum: [items, orders, deposits, batches, cash_flow, settings, manage_admins]
    AdminUserPublic:
      type: object
      required: [id, username, display_name, role, permissions]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/AdminPermission'
          description: Quyền hiệu lực (owner = đủ 7 quyền)
    AdminUser:
      allOf:
        - $ref: '#/components/schemas/AdminUserPublic'
        - type: object
          required: [is_active, created_at]
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
    CreateAdminRequest:
      type: object
      required: [username, password, display_name, role]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          format: password
          minLength: 6
        display_name:
          type: string
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/AdminPermission'
          description: Chỉ áp dụng cho staff
    UpdateAdminRequest:
      type: object
      properties:
        display_name:
          type: string
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/AdminPermission'
        password:
          type: string
          format: password
          minLength: 6
        is_active:
          type: boolean
...